Move RatingsFilter into FilterRibbon componenent.
