+++ src/components/AllMovies/FilterRibbon/FilterRibbon.js
@@ -2,9 +2,11 @@ import { useState } from "react"
 import './FilterRibbon.css'
 import crossIcon from '../../../images/cross.png'
 import PropTypes from 'prop-types';
-
+import RatingsFilter from '../RatingsFilter/RatingsFilter'
 
 const Searchbar = ({changeSearch, searchValue}) => {  
+
+  
   return (
     <div className="search-container">
       <input 
@@ -14,7 +16,7 @@ const Searchbar = ({changeSearch, searchValue}) => {
         type="text"
         onChange={event => changeSearch(event.target.value)}
         value={searchValue}
-      ></input>
+        ></input>
       <div className="cross-container">
         <img
           src={crossIcon}
@@ -27,9 +29,54 @@ const Searchbar = ({changeSearch, searchValue}) => {
 }
 
 const FilterRibbon = ({changeSearch, searchValue}) => {
+  const items = [
+    {
+      id: 1,
+      value: '1.0 +'
+    }, 
+    {
+      id: 2,
+      value: '2.0 +'
+    },  
+    {
+      id: 3,
+      value: '3.0 +'
+    }, 
+    {
+      id: 4,
+      value: '4.0 +'
+    }, 
+    {
+      id: 5,
+      value: '5.0 +'
+    }, 
+    {
+      id: 6,
+      value: '6.0 +'
+    },  
+    {
+      id: 7,
+      value: '7.0 +'
+    }, 
+    {
+      id: 8,
+      value: '8.0 +'
+    }, 
+    {
+      id: 9,
+      value: '9.0 +'
+    },
+    {
+      id: 10,
+      value: '10.0 +'
+    }
+  ];
+
   return (
-    <div className="all-movies-filter-ribbon">
+    <div className="all-movies-filter-ribbon"><>
       <Searchbar changeSearch={changeSearch} searchValue={searchValue}/>
+      <RatingsFilter items={items} />
+    </>
     </div>
   )
 }
