+++ src/components/AllMovies/RatingsFilter/RatingsFilter.css
@@ -6,3 +6,4 @@
 .dropdown-header {
   background-image: none;
 }
+
